incorrect page and size response

+++ backend/db/migrations/20240221082342-create-spot.js
@@ -65,26 +65,11 @@ module.exports = {
 
       },
       price: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL,
         allowNull: false,
 
       },
-      // numReviews: {
-      //   type: Sequelize.INTEGER,
-      //   allowNull: false 
-      // },
-
-      // avgStarRating: {
-      //   type: Sequelize.DECIMAL,
-      //   allowNull: false,
-
-      // },
-
-      // previewImage: {
-      //   type: Sequelize.STRING,
-      //   allowNull: false,
-
-      // },
+     
       createdAt: {
         allowNull: false,
         type: Sequelize.DATE,
